@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.1.1.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<version>3.0.5</version>
+		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>guru.springframework</groupId>
 	<artifactId>mb2g-spring-boot</artifactId>
@@ -15,7 +16,7 @@
 	<description>Example Project fpr Spring Boot</description>
 
 	<properties>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 	</properties>
 
 	<dependencies>
@@ -52,11 +53,11 @@
 				<artifactId>spring-boot-maven-plugin</artifactId>
 				<executions>
 					<execution>
-                        <id>pre-it</id>
-                        <goals>
+						<id>pre-it</id>
+						<goals>
 							<goal>start</goal>
 						</goals>
-                    </execution>
+					</execution>
 					<execution>
 						<id>post-it</id>
 						<goals>
