@@ -1,11 +1,11 @@
 package guru.springframework.springboot;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @SpringBootTest
 public class Mb2gSpringBootApplicationTests {
 
@@ -14,4 +14,3 @@ public class Mb2gSpringBootApplicationTests {
 	}
 
 }
-
